Fix totality warnings.
Unfortunatly, the function `flatten`, a where clause of `group` is no
longer being reported as total. Point-free style coding /is to blame/.

This commit removes the style to ensure wl-pprint is considered total.

Honestly, I think Idris should still see the function as total when
using point-free style, but I don't know what has changed in Idris for
it to complain.